+++ feincms3_sites/__init__.py
@@ -1,4 +1,2 @@
 VERSION = (0, 13, 1)
 __version__ = ".".join(map(str, VERSION))
-
-default_app_config = "feincms3_sites.apps.SitesAppConfig"