+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Change log
 
 - Add DefaultLanguageListFilter to limit choices to settings.LANGUAGE
 - Add default list_filter to SiteAdmin (is_active, host, default_language)
-- Added Python 3.10, Django 4.0 to the CI.
+- Added Python 3.10, Django 4.0 and 4.1 to the CI.
 - Raised the requirements to Python >= 3.8, Django >= 3.2.